Extract PermissionLifetimeOptions to dedicated file.
