+++ components/permissions/BUILD.gn
@@ -1,5 +1,7 @@
 source_set("permissions") {
   sources = [
+    "permission_lifetime_options.cc",
+    "permission_lifetime_options.h",
     "permission_lifetime_utils.cc",
     "permission_lifetime_utils.h",
   ]
