+++ components/permissions/permission_lifetime_utils.h
@@ -11,24 +11,9 @@
 #include "base/optional.h"
 #include "base/strings/string16.h"
 #include "base/time/time.h"
+#include "brave/components/permissions/permission_lifetime_options.h"
 
 namespace permissions {
-
-struct PermissionLifetimeOption {
-  PermissionLifetimeOption(base::string16 label,
-                           base::Optional<base::TimeDelta> lifetime);
-  PermissionLifetimeOption(const PermissionLifetimeOption&);
-  PermissionLifetimeOption& operator=(const PermissionLifetimeOption&);
-  PermissionLifetimeOption(PermissionLifetimeOption&&) noexcept;
-  PermissionLifetimeOption& operator=(PermissionLifetimeOption&&) noexcept;
-  ~PermissionLifetimeOption();
-
-  // Text visible to the user.
-  base::string16 label;
-  // If not set, lifetime will not be controlled (i.e. permanent). If set to
-  // base::TimeDelta(), permission should be alive until eTLD+1 is closed.
-  base::Optional<base::TimeDelta> lifetime;
-};
 
 // Returns pre-configured permission lifetime options.
 std::vector<PermissionLifetimeOption> CreatePermissionLifetimeOptions();
